requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
firebase-admin==6.6.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.10.0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
openpyxl==3.1.5
xlrd==2.0.1
python-dateutil==2.9.0.post0
google-cloud-storage==2.19.0
google-cloud-tasks==2.16.5
httpx==0.27.2
pytest==8.3.3
Jinja2==3.1.4
Pillow==10.4.0
WeasyPrint==62.3
qrcode==7.4.2
